Add vdev contraction to AnyRAID
Contraction is the last piece of the puzzle for allowing AnyRAID to be
flexible in terms of allowing users to add and use storage to their
vdevs as they wish. Thus far, it has not been possible to shrink the
logical size of a vdev in ZFS; mirrors can be detached, but that doesn't
actually reduce the available space in the pool, just the amount of
parity in the mirror. Device removal removes a whole top-level vdev, but
doesn't shrink an individual vdev. Contraction is different; it actually
shrinks the top-level vdev but leaves it in place.

This works by taking all the tiles on the leaf vdev being removed and
moving them to other devices in the anyraid vdev. This does impose some
restrictions, since if we cannot find a way to move all the tiles to
other leaf vdevs, the contraction cannot happen without reducing
redundancy. This plan generation step works similarly to rebalance, and
could be improved in the future with more advanced algorithms.

The second an third phases of contraction are also similar to rebalance,
and much of the code is shared between the two. The final phase is the
actual shrinking of the top-level vdev, which only takes a single
txg. In order to shrink a top-level vdev we need to ensure that no data
is present in the metaslabs that will be removed when the asize is
reduced; if there is anything allocated in those regions we fail to
contract, even if that data could be moved. Future work may allow for a
device-removal-like remapping of that data before contraction occurs.

Sponsored-by: Eshtek, creators of HexOS
Sponsored-by: Klara, Inc.
Contributions-by: JT Pennington <jt.pennington@klarasystems.com>

+++ include/sys/spa_impl.h
@@ -334,7 +334,7 @@ struct spa {
 	kthread_t	*spa_async_thread;	/* thread doing async task */
 	int		spa_async_suspended;	/* async tasks suspended */
 	kcondvar_t	spa_async_cv;		/* wait for thread_exit() */
-	uint16_t	spa_async_tasks;	/* async task mask */
+	uint32_t	spa_async_tasks;	/* async task mask */
 	uint64_t	spa_missing_tvds;	/* unopenable tvds on load */
 	uint64_t	spa_missing_tvds_allowed; /* allow loading spa? */
